pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "brand-qa-assistant"
version = "0.1.0"
description = "Web app that performs brand compliance reviews using Azure AI Foundry"
requires-python = ">=3.12"
dependencies = [
    # Azure AI
    "azure-ai-agents>=1.1.0",
    "azure-identity>=1.21.0",
    "azure-storage-blob>=12.24.0",
    # Web server
    "aiohttp>=3.11.0",
    # Screenshot
    "playwright>=1.50.0",
    # Config
    "pydantic-settings>=2.7.0",
    # Retry
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
